+++ beecube-module-paimai/beecube-module-paimai-biz/src/main/java/cn/winkt/modules/paimai/service/JobService.java
@@ -212,7 +212,7 @@ public class JobService {
                     commission = BigDecimal.valueOf(goods.getCommission()).divide(BigDecimal.valueOf(100), 4, RoundingMode.CEILING);
                 }
 
-                BigDecimal price = BigDecimal.valueOf(maxOfferRow.getPrice());
+                BigDecimal price = maxOfferRow.getPrice();
                 BigDecimal newPrice = price.multiply(commission).add(price).setScale(2, RoundingMode.CEILING);
                 //生成成交订单
                 GoodsOrder goodsOrder = new GoodsOrder();
